[IMP] Moved class AlchemyEncoder into blueprints namespace
